LayoutTests/editing/caret/ios/caret-color-in-nested-editable-containers-expected.txt


This test verifies exercises the behavior of caret-color in nested editable containers. To manually run the test, focus the red box and verify that the caret color is red.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS caretColor is "rgb(255, 0, 0)"
PASS successfullyParsed is true

TEST COMPLETE
